+++ engine/graphics/source/graphics/material/standard-material.hpp
@@ -45,6 +45,7 @@ namespace april::graphics
 
         // IMaterial interface
         auto getType() const -> generated::MaterialType override;
+        auto getTypeName() const -> std::string override;
         auto writeData(generated::StandardMaterialData& data) const -> void override;
         auto getTypeConformances() const -> TypeConformanceList override;
         auto bindTextures(ShaderVariable& var) const -> void override;
@@ -52,6 +53,8 @@ namespace april::graphics
         auto getFlags() const -> uint32_t override;
         auto setDoubleSided(bool doubleSided) -> void override;
         auto isDoubleSided() const -> bool override;
+        auto serializeParameters(nlohmann::json& outJson) const -> void override;
+        auto deserializeParameters(nlohmann::json const& inJson) -> bool override;
 
         auto setDescriptorHandles(
             uint32_t baseColorTextureHandle,