+++ engine/graphics/source/graphics/material/unlit-material.hpp
@@ -12,6 +12,7 @@ namespace april::graphics
         ~UnlitMaterial() override = default;
 
         auto getType() const -> generated::MaterialType override;
+        auto getTypeName() const -> std::string override;
         auto writeData(generated::StandardMaterialData& data) const -> void override;
         auto getTypeConformances() const -> TypeConformanceList override;
         auto bindTextures(ShaderVariable& var) const -> void override;
@@ -19,6 +20,8 @@ namespace april::graphics
         auto getFlags() const -> uint32_t override;
         auto setDoubleSided(bool doubleSided) -> void override;
         auto isDoubleSided() const -> bool override;
+        auto serializeParameters(nlohmann::json& outJson) const -> void override;
+        auto deserializeParameters(nlohmann::json const& inJson) -> bool override;
 
         float4 color{1.0f, 1.0f, 1.0f, 1.0f};
         float3 emissive{0.0f, 0.0f, 0.0f};